# apps/api/api/openapi/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Follytics API service",
        "title": "Follytics API",
        "termsOfService": "https://follytics.localhost/terms/",
        "contact": {
            "name": "Follytics Support",
            "url": "http://follytics.localhost/support",
            "email": "support@follytics.localhost"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "host": "api.follytics.localhost",
    "basePath": "/",
    "paths": {
        "/users/login-events": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new user if they do not exist or updates an existing user upon login",
                "tags": [
                    "User"
                ],
                "summary": "Track user login",
                "operationId": "trackLogin",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPResponse-response_User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns details of the currently authenticated account",
                "tags": [
                    "User"
                ],
                "summary": "Retrieve the current authenticated user",
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPResponse-response_User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{id}/follow-events": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a paginated list of follow/unfollow events for the specified user ID",
                "tags": [
                    "User"
                ],
                "summary": "Retrieve follow events for a user",
                "operationId": "getUserFollowEvents",
                "parameters": [
                    {
                        "type": "string",
                        "format": "int",
                        "description": "Page number for pagination (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "int",
                        "description": "Number of results per page (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID to retrieve followers for",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.HTTPResponse-array_response_EventWithUserReference"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "pagination": {
                                            "$ref": "#/definitions/pagination.Metadata"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{id}/followers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a paginated list of followers for the specified user ID",
                "tags": [
                    "User"
                ],
                "summary": "Retrieve followers for a user",
                "operationId": "getUserFollowers",
                "parameters": [
                    {
                        "type": "string",
                        "format": "int",
                        "description": "Page number for pagination (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "int",
                        "description": "Number of results per page (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID to retrieve followers for",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.HTTPResponse-array_response_User"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "pagination": {
                                            "$ref": "#/definitions/pagination.Metadata"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{id}/followers/timeline": {
            "get": {
                "description": "Returns a timeline of user followers for the specified user ID",
                "tags": [
                    "User"
                ],
                "summary": "Retrieve followers timeline for a user",
                "operationId": "getUserFollowersTimeline",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID to retrieve timeline for",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPResponse-response_FollowersTimeline"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "pagination.Metadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "totalItems": {
                    "type": "integer",
                    "example": 100
                },
                "totalPages": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "response.EventWithUserReference": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "type"
            ],
            "properties": {
                "createdAt": {
                    "description": "nolint: lll",
                    "type": "string",
                    "example": "1970-01-01T00:00:00.000+04:00"
                },
                "id": {
                    "type": "string",
                    "example": "01955908-d43b-7900-8f5c-5faa67dab4d3"
                },
                "type": {
                    "type": "string",
                    "example": "FOLLOW"
                },
                "user": {
                    "$ref": "#/definitions/response.UserForEventReference"
                }
            }
        },
        "response.FollowersTimeline": {
            "type": "object",
            "required": [
                "timeline",
                "user"
            ],
            "properties": {
                "timeline": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.FollowersTimelineItem"
                    }
                },
                "user": {
                    "$ref": "#/definitions/response.UserForTimeline"
                }
            }
        },
        "response.FollowersTimelineItem": {
            "type": "object",
            "required": [
                "date",
                "follows",
                "total",
                "unfollows"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "example": "1970-01-01T00:00:00.000+04:00"
                },
                "follows": {
                    "type": "integer",
                    "example": 100
                },
                "total": {
                    "type": "integer",
                    "example": 1500
                },
                "unfollows": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "response.HTTPError": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "UNKNOWN"
                },
                "message": {
                    "type": "string",
                    "example": "some unknown error"
                }
            }
        },
        "response.HTTPResponse-array_response_EventWithUserReference": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.EventWithUserReference"
                    }
                }
            }
        },
        "response.HTTPResponse-array_response_User": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.User"
                    }
                }
            }
        },
        "response.HTTPResponse-response_FollowersTimeline": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.FollowersTimeline"
                }
            }
        },
        "response.HTTPResponse-response_User": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.User"
                }
            }
        },
        "response.User": {
            "type": "object",
            "required": [
                "avatar",
                "createdAt",
                "email",
                "id",
                "name",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://avatars.githubusercontent.com/u/123456"
                },
                "createdAt": {
                    "type": "string",
                    "example": "1970-01-01T00:00:00.000+04:00"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "01955908-d43b-7900-8f5c-5faa67dab4d3"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "username": {
                    "type": "string",
                    "example": "john.doe"
                }
            }
        },
        "response.UserForEventReference": {
            "type": "object",
            "required": [
                "avatar",
                "id",
                "name",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://avatars.githubusercontent.com/u/123456"
                },
                "id": {
                    "type": "string",
                    "example": "01955908-d43b-7900-8f5c-5faa67dab4d3"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "username": {
                    "type": "string",
                    "example": "john.doe"
                }
            }
        },
        "response.UserForTimeline": {
            "type": "object",
            "required": [
                "avatar",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "data:image/jpeg;base64,JCQgKDBQND="
                },
                "username": {
                    "type": "string",
                    "example": "john.doe"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}